fix(lint): resolve goconst, gosec, lll issues across all services
Resolves all remaining golangci-lint issues (goconst, gosec, lll, revive) across all 7 services and shared packages.

+++ apps/budget-svc/internal/infrastructure/api/grpc_handler_test.go
@@ -1,3 +1,4 @@
+//nolint:goconst // test file - repeated strings acceptable for readability
 package api_test
 
 import (
@@ -19,7 +20,7 @@ import (
 	budgetv1 "github.com/aureum/proto/gen/budget/budgetv1"
 )
 
-// Mock application service
+// Mock application service.
 type mockAppService struct {
 	mock.Mock
 }
@@ -462,7 +463,7 @@ func TestGRPCHandler_GetBudgetSummary_Success(t *testing.T) {
 	assert.Equal(t, summary.TotalSpent, protoResp.TotalSpent)
 	assert.Equal(t, summary.Remaining, protoResp.Remaining)
 	assert.InDelta(t, summary.UsagePercent, protoResp.UsagePercentage, 0.01)
-	assert.Equal(t, int32(summary.CategoryCount), protoResp.CategoryCount)
+	assert.Equal(t, summary.CategoryCount, protoResp.CategoryCount)
 	assert.Len(t, protoResp.Categories, 1)
 	assert.Equal(t, summary.Categories[0].CategoryID, protoResp.Categories[0].CategoryId)
 
@@ -487,7 +488,7 @@ func TestGRPCHandler_GetBudgetSummary_NotFound(t *testing.T) {
 	assert.Equal(t, codes.NotFound, status.Code(err))
 }
 
-// Helper functions for pointer creation
+// Helper functions for pointer creation.
 func ptr(s string) *string {
 	return &s
 }
